Add files via upload
The project works as a windows application application. In short, the project is a system that allows you to clean up duplicate data if there are any by uploading excel files, selecting and listing the relevant headings. The cleaned data can be re-downloaded as an excel file. The application is written in python.